# ui3/src/generated/catalyst/openapi/communities.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "catalyrst-social-service",
    "description": "community REST surface",
    "license": {
      "name": ""
    },
    "version": "0.15.3"
  },
  "paths": {
    "/federation/communities/changes": {
      "get": {
        "tags": [
          "federation"
        ],
        "operationId": "changes",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/federation/communities/content/{hash}": {
      "get": {
        "tags": [
          "federation"
        ],
        "operationId": "get_content",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "content bytes",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorBody": {
        "type": "object",
        "required": [
          "ok",
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "federation_adr": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          }
        }
      }
    }
  }
}